feat: clean code in order to separate routes, controllers and schemas
